fix(npm): use correct node version when running npm or npx
- refactor src folder structure to apply to golang-standards project-layout
