+++ .github/workflows/tests.yml
@@ -1,11 +1,6 @@
 name: run-tests
 
 on:
-  push:
-    branches:
-      - '**'
-    tags-ignore:
-      - '**'
   pull_request:
 
 jobs:
